+++ include/ambition/7kaaInterface/draw.hh
@@ -25,7 +25,9 @@
 
 #pragma once
 
+class Button;
 class Button3D;
+class ButtonCustom;
 class Firm;
 class FirmHarbor;
 class FirmWar;
@@ -122,6 +124,26 @@ void buttonCost(
   const int right,
   const int bottom,
   const int verticalAlignment = 1
+);
+
+void buttonKeybind(
+  const unsigned int keyCode,
+  const int left,
+  const int top,
+  const int right,
+  const int bottom
+);
+void buttonKeybind(
+  const unsigned int keyCode,
+  const Button& button
+);
+void buttonKeybind(
+  const unsigned int keyCode,
+  const Button3D& button
+);
+void buttonKeybind(
+  const unsigned int keyCode,
+  const ButtonCustom& button
 );
 
 short calculateAnimatedLinePhase(