+++ CHANGELOG.md
@@ -16,6 +16,9 @@ speaking, considered to be the player's experience.
 - Show war machine and ship construction costs on War Factory and Harbour build
   menus.
 - Show total production queue count and cost for War Factories and Harbours.
+- Show keybinds at the bottom-left of buttons that have them.
+- Add a keybind for cancelling menus (defaults to Esc), settable in config.txt
+  with `bindkey = KEYEVENT_CANCEL,Escape`.
 
 ### Changed